website/requirements.txt
django
django-registration
django-crispy-forms
psycopg2
pillow


